requirements.txt
Django==4.2.5
Pillow==10.0.1
djangorestframework==3.14.0